requirements.txt
requests==2.31.0
pandas==2.1.4
sqlalchemy==2.0.23
python-telegram-bot==20.6
Flask==3.0.0